requirements.txt
autopep8
fire
flake8
flake8-import-order
playwright
python-dotenv
pytz